Added territory buffs, changed some var names, and created a basic listener
Did a bunch of work on this one. Firstly, I added a bit of code that allows
the bot to calculate defensive buffs for territories based on the buildings
within them. Added a couple of buildings which give territory buffs as
well. I also wrote a brief, bare-bones prototype of a listener that should,
once finished, allow me to implement an API for Reo and Chromabot that lets
it check for territory defensive buffs.

Signed-off-by: Eliminioa <jmsboggs@gmail.com>